requirements.txt
django==2.2.6
django-choices==1.7.0
djangorestframework==3.10.3
